refactor(network): trim intercept stack dead code and duplication
Remove unused CyIntercept extra-target handling, dead strip-internal-headers
module, duplicate main-tab CSP in proxy, and consolidate driver conversion
and adapter test boilerplate for net-negative LOC.
