+++ packages/net-stubbing/lib/index.ts
@@ -1,1 +1,1 @@
-export { CyIntercept, INTERCEPT_HEADERS } from './cy-intercept'
+export { CyIntercept } from './cy-intercept'
