+++ packages/net-stubbing/lib/types.ts
@@ -1,3 +1,3 @@
 export * from './external-types'
 
-export * from './driver-intercept'
+export * from './driver-http-conversion'
