--- packages/network-interception/lib/ports/driving-ports.ts
@@ -1,1 +1,0 @@
-export * from './http-interception'
